Support instantiating server by ID
This is so if you already know the ID (maybe b/c the server was created
via a Heat template, for example), you can get a nice representation of
it. Updates the signature to require either id or lease (prefers id). If
lease is provided, it creates the server as before. The only required
arg is now the keystone session.