Support all Keystone auth methods
The prior version of the context helper only could handle either
password or token login. We however need to support at least OIDC
methods for other applications. This commit attempts to solve  some
long-standing headaches with Keystone session management, hopefully
making it more predictable and less annoying for users.

It works by dealing with hidden oslo.config abstractions under the hood.
The chi "context" is a collection of authentication parameters and some
additional parameters specific to the operation of the python-chi
library.

- Any auth method is supported. By default, the auth method is inferred
  from the value of OS_AUTH_TYPE
- Auth parameters are defaulted from their OS_* env setting, where the
  rule is the env name is the parameter name uppercased, e.g., the
  "identity_provider" param is set by OS_IDENTITY_PROVIDER.
- Callers cannot get or set invalid or unsupported options and will get an
  error if trying.
- Some auth providers support the same option, e.g. "project_id". When
  setting the project_id parameter, this parameter is configured for all
  authentication plugins that support it.
- Callers can set parameters for auth plugins that are currently not set
  to be active (via the "auth_type" parameter.) This allows setting
  params in any order, e.g. first setting "identity_provider" (a
  v3oidcaccesstoken param) and then setting "auth_type" to
  "v3oidcaccesstoken".