Fixup automatic access token refresh logic
Overriding only Session.request doesn't catch a few other use-cases,
notably looking up an endpoint from a Keystone service catalog. Since
this can happen _before_ request is called, it will fail. So, we have to
do something else: this commit updates the access_token on the auth
object to actually be a dynamically-bound @property, which can detect if
it needs to refresh synchronously.