Pick first flavor if none specified
This updates create_server to be a bit more agnostic about whether
it's running against a KVM cloud or not. Instead of assuming the
baremetal flavor is available, we just try to enumerate all flavors
and pick the first one found. For baremetal sites, this should be
the baremetal flavor; for KVM, it should be the smallest flavor.