Migrate legacy `chi.Server` objects to own a `openstack.compute.v2.server.Server` object instead of a `novaclient.v2.servers.Server` object
Users attempting to use the legacy object-oriented CHI API were encountering `AttributeError`s when deleting instances because various OpenStack calls were expecting to receive an OpenStack `Server` object, but were instead receiving novaclient's, which does not implement the same interface (`openstack.resource.Resource`). To remedy this issue, the object has been swapped out for the correct type.

This refactor slots in nicely to the existing code because basically all existing usages of it were expecting OpenStack `Server`s and we were merely lucking out due to duck typing. As such, this refactor should prevent similar bugs from occuring in the future, until the legacy interface is eventually removed.

Another small change was made in the `chi.Server.delete` method as well, swapping out a call to `wait_for_server` for `wait_for_delete`, which accounts for the resource being inaccessible once it is deleted. Without this change, the wait call was very likely to throw if its retry interval did not happen to sync up with the actual `DELETE` operation taking place. Now, the only time `delete` is expected to throw is if it is called on a resource which already didn't exist at call-time.

Fixes #12.