Fixup handling of empty args, compat w/ v2
This defaults all requests to platform version 2, since we are now
rolled over to that.

It also fixes up some default arg handling. The Zun client will not
send an arg if it is not on kwargs, and the API will interpret that
as the arg is missing, which is OK for optional args. Importantly, some
args MUST be missing (exposed_ports when security_groups are defined,
for example.) This updates the handling to only document some helpful
args but not actually include them on the function signature, meaning
they will ride along w/ kwargs, if defined.