Handle mixed leases when searching for reservations with counts (#35)
In `chi.lease.get_node_reservation`, if a `count` is provided, the lookup will fail because it searches all reservations for the values of `min_count` and `max_count`. Floating IP reservations have only an `amount`, so the comparison fails.

This change simply uses an invalid count for any reservations which have un-bounded counts.