Added custom exceptions
Raised when argument is not valid. These errors might be fixed by checking hardware catalog or documentation. Examples where this might be seen are:
- Site name is not valid
- Node type is not valid
- e.g. Resource does not exist

Raised when a request has valid arguments, but the resources are being used incorrectly, or can’t be used as requested. This type of error might depend on the time the notebook is run, due to the shared nature of the testbed.
Examples:
- Nodes matching filters (e.g. node_type) are unavailable
- Cannot allocate FIP
- Allocation expires soon
- Allocation has insufficient SUs for request

Raised when an error occurs with some Chameleon resource.
For example, if your node is having hardware issues, and so fails to provision, this will be raised.

Replaced thrown exceptions with their appropriate custom exception
accross all modules.