add unittest for container upload
this tests the case where we upload a single file to a running
container. To be a good test, we need to also test the case where
the source is a directory, and cases where the destination is a file,
and not a directory.

Proper testing requires a refactor of the upload/download code to take
a file descriptor instead of a path, and separation of tar/encoding
into a separate method.