add unittests for container file upload/download
in chi@edge, e.g. zun-api backed by kubernetes, file upload/download
is fairly brittle. The API implementaton assumes that the payload is
a tar file, in base64 encoding. The tar file may include multiple files,
but the destination of the upload/download is always a directory, even
in the case that a single file is sent.

As this is unintuitive, and frequent errors happen due to file encoding,
these tests try to illustrate the "known working" behavior.